implement a spanish game class
feature/spanishgame

This commit has a spanish version of the game that extends the existing game.
The only difference is how the deck is built. There is a commented out method
and a not commented out method. This is because to run the tests I needed
to have a version of the code that compiled. The commented out version
will be the final version, but it won't work until there is code for the
spanish suits.