Added EnglishGame subclass which extends from Game class.
The only game functions I found necessary to implement separately were the English/Spanish Game constructors and the buildDeck implementations. I was testing the whole time, and even edited tests for the Card and Game classes, as they had not yet specified English or Spanish Game when calling the constructor.