Commit -- EnglishGame subclass
I moved the two functions which I thought needed to be implemented seperately, constructor and buildDeck functions, into a new EnglishGame subclass.
Along the way I created tests for the functions, as well as edited testCard and testGame files because they were refering to the Game object and not an EnglishGame or SpanishGame object.